+++ tests/PhysicalQuantity/PowerTest.php
@@ -49,4 +49,24 @@ class PowerTest extends AbstractPhysicalQuantityTestCase
         $quantity = new Power(1, 'kW');
         $this->assertEquals(1000, $quantity->toUnit('W'));
     }
+
+    public function testToDecibelsMilliWatt()
+    {
+        $quantity = new Power(0.01, 'mW');
+        $this->assertEquals(-20, $quantity->toUnit('dBm'));
+        $quantity = new Power(1, 'mW');
+        $this->assertEquals(0, $quantity->toUnit('dBm'));
+        $quantity = new Power(10, 'kW');
+        $this->assertEquals(70, $quantity->toUnit('dBm'));
+    }
+
+    public function testFromDecibelsMilliWatt()
+    {
+        $quantity = new Power(-20, 'dBm');
+        $this->assertEquals(0.00001, $quantity->toUnit('W'));
+        $quantity = new Power(0, 'dBm');
+        $this->assertEquals(0.001, $quantity->toUnit('W'));
+        $quantity = new Power(70, 'dBm');
+        $this->assertEquals(10000, $quantity->toUnit('W'));
+    }
 }